Implementation of q_delete_mid
In q_delete_mid(), we use pointer to pointer method to remove node from
a linked list. we use fast and slow (which is indir in code) pointer to
find the middle node. In every step, fast go forward for two nodes and
indir go forward for one node. When fast arrive the end of queue, indir
will be the middle of queue.