Implement q_delete_dup
In this function, we need to delete duplicated nodes. So we use
list_for_each_entry_safe() macro so that we can iterate all nodes in
queue and delete the node by using the "entry" iterator. Then, because
the "entry" is just the previous node of "safe" node, we can compare
their value to find the duplicated nodes. We use q_release_element to
free this linked list node and its entry structure.