Implementation of q_swap
The concept of swap nodes is we treated a node as a temporary head. And
we decide to swap the two node next to this temporary head. First, we
target the second node next to the head. Then we use list_move() API to
move it to next of this temporary head.