Implementation of q_reverse, q_reverseK
We use pointer to pointer method to implement q_reverse and q_reverseK.
Using pointer to pointer can save many operation of queue. We point
**indir to the address of the next in the first node. So every time we
move the node to the head of queue, the pointer **indir always point to
the next node which next to the original first node. We keep moving
nodes to head of queue until the next of orignal first node point to the
head of queue.

q_reverseK has some difference. In order to reverse K nodes, we have to
change the place we move the nodes to. So there is a temporary head that
we can move node to the next of the temporary head.