Fix Segfault and loop error in q_delete_mid
Orginally, we assigned head to fast pointer, but it will stop because the
condition of for loop is "fast != head". So we change **indir and fast
to point to the next node of head so we can step into the for loop
normally.

There has segmentation fault while we try to free the pointer to
pointer. So we have to use another pointer to record the address in the
**indir and free the middle node from the pointer.