Implementation of q_sort
Implement q_sort with merge sort. First, we break the link which
last element connect to head. Next step, we start doing merge sort and
also break the link from middle node because that easier to determine
the end of list. In the merge_sort(), we use fast and slow pointer
method to find the middle node. After recursion finished, we merge list
with  merge_two_lists(). When the list is sorted, it has to reconstruct
the connection we breaked before.
merge_two_lists() use pointer to pointer that point to the head of
result list. We can add a node to this list without extra operation. In
the end of merge_two_lists function, we concatenate the remaining nodes
to result list by using bitwise OR because one of "L" and "R" is NULL.
The result of bitwise OR is the pointer of "L" or "R".