Fix wrong implementation in q_delete_dup
q_delete_dup() needs to delete all duplicated nodes. The function we
implemented before would leave one of duplicated nodes. We delete the
last duplicated node by using another variable "dup" to record. If "dup"
is set, it means that these nodes are duplicated. When it runs to the
last duplicate node, it would be deleted by last if-statement.